+++ aws/content-site/main.tf
@@ -64,7 +64,7 @@ resource "aws_s3_bucket" "logs" {
   tags = {
     Purpose         = "Log bucket for static site ${var.domain}"
     Terraform       = true
-    TerraformModule = "github.com/halostatue/terraform-modules//aws/content-site@v3.0.0"
+    TerraformModule = "github.com/halostatue/terraform-modules//aws/content-site@v3.1.1"
   }
 }
 
@@ -107,7 +107,7 @@ resource "aws_s3_bucket" "bucket" {
   tags = {
     Purpose         = "Bucket for static site ${var.domain}"
     Terraform       = true
-    TerraformModule = "github.com/halostatue/terraform-modules//aws/content-site@v3.0.0"
+    TerraformModule = "github.com/halostatue/terraform-modules//aws/content-site@v3.1.1"
   }
 
   lifecycle {
@@ -125,7 +125,7 @@ resource "aws_iam_user" "publisher" {
   tags = {
     Purpose         = "Publishing user for ${var.domain}"
     Terraform       = true
-    TerraformModule = "github.com/halostatue/terraform-modules//aws/content-site@v3.0.0"
+    TerraformModule = "github.com/halostatue/terraform-modules//aws/content-site@v3.1.1"
   }
 }
 
@@ -177,7 +177,7 @@ resource "aws_iam_policy" "publisher" {
   tags = {
     Purpose         = "Publishing user policy for ${var.domain}"
     Terraform       = true
-    TerraformModule = "github.com/halostatue/terraform-modules//aws/content-site@v3.0.0"
+    TerraformModule = "github.com/halostatue/terraform-modules//aws/content-site@v3.1.1"
   }
 }
 
@@ -267,6 +267,6 @@ resource "aws_cloudfront_distribution" "content" {
   tags = {
     Purpose         = "Cloudfront distribution for ${var.domain}"
     Terraform       = true
-    TerraformModule = "github.com/halostatue/terraform-modules//aws/content-site@v3.0.0"
+    TerraformModule = "github.com/halostatue/terraform-modules//aws/content-site@v3.1.1"
   }
 }